+++ examples/minimal_project/build.py
@@ -8,11 +8,8 @@ project = sbuildr.Project()
 # Full file paths are only required in cases where a partial path would be ambiguous.
 libmath = project.library("math", sources=["factorial.cpp", "fibonacci.cpp"], libs=["stdc++"])
 
-# Set the installation location for the library created above to /usr/local/lib.
-project.install(libmath, path=os.path.join("/", "usr", "local", "lib"))
-
-# Set the installation location for the public header to /usr/local/include.
-project.install("math.hpp", path=os.path.join("/", "usr", "local", "include"))
+# Specify that math.hpp is part of the public API for this library.
+project.interfaces(["math.hpp"])
 
 # Specify a test for the project using the test.cpp source file. The resulting executable will
 # be linked against the library created above.